Complain if UTF-16 documents try to switch to UTF-8
That is, if a document has a UTF-16 byte order mark, complain if the
XML header specifies UTF-8. The XML header can specify which of
several versions of the current encoding the document uses, but it
can't change the encoding completely.

The actual issue here is dumb XML serializers that print "UTF-8"
whatever the actual encoding may be.